add dialog for delete film

+++ app/src/main/java/com/example/fdbsilverfilm/adapter/FilmAdapter.kt
@@ -1,5 +1,6 @@
 package com.example.fdbsilverfilm.adapter
 
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.view.LayoutInflater
@@ -9,7 +10,6 @@ import android.widget.ImageButton
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.fdbsilverfilm.R
-import com.example.fdbsilverfilm.manager.DatabaseManager
 import com.example.fdbsilverfilm.manager.SharedPreferencesManager
 import com.example.fdbsilverfilm.model.Film
 import com.example.fdbsilverfilm.view.MainActivity
@@ -49,7 +49,20 @@ class FilmAdapter(
 
         holder.deleteButton.setOnClickListener {
             films[position].let {
-              filmViewModel.deleteFilm(it)
+
+                val builder = AlertDialog.Builder(context)
+                builder.setTitle(context.getString(R.string.delete_film))
+                builder.setMessage(context.getString(R.string.ask_delete_film))
+
+                builder.setPositiveButton(android.R.string.yes) { _, _ ->
+                    filmViewModel.deleteFilm(it)
+                }
+
+                builder.setNegativeButton(android.R.string.no) { dialog, _ ->
+                    dialog.dismiss()
+                }
+
+                builder.show()
             }
         }
     }
